Item Destroy Development Updated 1
コードレビューの内容に基づき、コードを修正しました。

+++ app/controllers/items_controller.rb
@@ -36,6 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    redirect_to root_path unless item.user_id == current_user.id
     item.destroy
     redirect_to root_path
   end
